-created 2 dictionaries for shape [CLLocation:CLLocation] -created CLLocation Type to set as key in shape dictionary -fixed intersection coordinates issue
