+++ Overrun/Run.swift
@@ -153,98 +153,5 @@ class Run: NSObject {
             totalDistance += (runLocations.last?.distance(from: lastKnownLocation))!
         }
         lastKnownLocation = runLocations.last!
-    }
-    
-    func checkShapeIntersection(existingRuns: [Run]) -> [CLLocationCoordinate2D]{
-        
-        var intersectingCoorArray = [CLLocationCoordinate2D]()
-        
-        for specificRun in existingRuns{
-            
-            var indexNewP1 = 0
-            var indexNewP2 = 1
-            
-             for _ in 0..<(smartArray.count) - 1{
-                
-                if indexNewP2 == smartArray.count - 1{
-                    
-                    indexNewP1 = 0
-                }
-                
-                var indexPulledP3 = 0
-                var indexPulledP4 = 1
-                
-                for _ in 0..<(specificRun.coorArray?.count)! - 1{
-                    
-                    if indexPulledP4 == (specificRun.coorArray?.count)! - 1{
-                        
-                        indexPulledP3 = 0
-                    }
-                    
-                    let newP1 = self.smartArray[indexNewP1].coordinate
-                    let newP2 = self.smartArray[indexNewP2].coordinate
-                    
-                    let pulledP3 = specificRun.coorArray?[indexPulledP3]
-                    let pulledP4 = specificRun.coorArray?[indexPulledP4]
-                
-                    let d1 = ((newP2.longitude) - (newP1.longitude))*((pulledP4?.latitude)! - (pulledP3?.latitude)!)
-                    let d2 = ((newP2.latitude) - (newP1.latitude))*((pulledP4?.longitude)! - (pulledP3?.longitude)!)
-                    let d = d1 - d2
-                    if (d == 0) {
-                        print("LINES ARE PARALLEL")
-                        
-                        indexPulledP3 += 1
-                        indexPulledP4 += 1
-                        print("P3: %d", indexPulledP3)
-                        print("P4: %d", indexPulledP4)
-                        
-                    } else {
-                        
-                        let u1 = ((pulledP3?.longitude)! - (newP1.longitude))*((pulledP4?.latitude)! - (pulledP3?.latitude)!)
-                        let u2 = ((pulledP3?.latitude)! - (newP1.latitude))*((pulledP4?.longitude)! - (pulledP3?.longitude)!)
-                        let u = (u1 - u2)/d
-                        
-                        let v1 = ((pulledP3?.longitude)! - (newP1.longitude))*((newP2.latitude) - (newP1.latitude))
-                        let v2 = ((pulledP3?.latitude)! - (newP1.latitude))*((newP2.longitude) - (newP1.longitude))
-                        let v = (v1 - v2)/d
-                        
-                        if (u < 0.0 || u > 1.0){
-                            print("INTERSECTION POINT NOT BETWEEN p1 and p2")
-                            indexPulledP3 += 1
-                            indexPulledP4 += 1
-                            print("P3: %d", indexPulledP3)
-                            print("P4: %d", indexPulledP4)
-                        
-                        } else if (v < 0.0 || v > 1.0){
-                            print("INTERSECTION POINT NOT BETWEEN p3 and p4")
-                            indexPulledP3 += 1
-                            indexPulledP4 += 1
-                            print("P3: %d", indexPulledP3)
-                            print("P4: %d", indexPulledP4)
-                            
-                        } else {
-                            
-                            let intersectingCoor = CLLocationCoordinate2D(latitude: ((newP1.latitude) + u * ((newP2.latitude) - (newP1.latitude))), longitude: ((newP1.longitude) + u * ((newP2.longitude) - (newP1.longitude))))
-                            print(intersectingCoor)
-                            indexPulledP3 += 1
-                            indexPulledP4 += 1
-                            print("P3: %d", indexPulledP3)
-                            print("P4: %d", indexPulledP4)
-                            intersectingCoorArray.append(intersectingCoor)
-                        }
-                    }
-                }
-                if indexNewP1 == smartArray.count{
-                    break
-                }
-                indexNewP1 += 1
-                indexNewP2 += 1
-                
-                print("P1: %d", indexNewP1)
-                print("P2: %d", indexNewP2)
-            }
-        }
-        print(intersectingCoorArray.count)
-        return intersectingCoorArray
-    }
+    }    
 }
