+++ Overrun/ViewController.swift
@@ -11,7 +11,7 @@ import FirebaseAuth
 import FirebaseDatabase
 import SceneKit
 
-class ViewController: UIViewController, GMSMapViewDelegate, LocationManagerDelegate, SignInDelegate {
+class ViewController: UIViewController, GMSMapViewDelegate, LocationManagerDelegate, SignInDelegate, RunManagerDelegate {
     
     @IBOutlet weak var PlayerAnimationView: UIView!
     @IBOutlet weak var distanceLabel: UILabel!
@@ -22,6 +22,7 @@ class ViewController: UIViewController, GMSMapViewDelegate, LocationManagerDeleg
     @IBOutlet var endRunButton: UIButton!
     
     let locationManager = LocationManager()
+    let runManager = RunManager()
     var mapView:GMSMapView!
     var centerCoordinate: CLLocationCoordinate2D?
     
@@ -34,7 +35,9 @@ class ViewController: UIViewController, GMSMapViewDelegate, LocationManagerDeleg
     
     override func viewDidLoad() {
         super.viewDidLoad()
-    
+        
+        runManager.delegate = self
+        
         // create a new scene
         let scene = SCNScene(named: "art.scnassets/ship.scn")!
         
@@ -119,54 +122,8 @@ class ViewController: UIViewController, GMSMapViewDelegate, LocationManagerDeleg
         
         distanceView.isHidden = true
         
-        pullRunsFromFirebase()
+        runManager.pullRunsFromFirebase()
         
-    }
-    
-    func pullRunsFromFirebase() {
-        
-        let shapeRef = FIRDatabase.database().reference()
-        shapeRef.child("Runs").observe(FIRDataEventType.value) { (shapeSnap: FIRDataSnapshot) in
-            
-            if let snapDict = shapeSnap.value as? NSDictionary{
-                
-                for snapRun in snapDict{
-                    
-                    let uID = snapRun.key as? String
-                    
-                    let run = snapRun.value as? NSDictionary
-                    
-                    var coordinateArray = [CLLocationCoordinate2D]()
-                    
-                    let runCoordinates = run?.value(forKey: "coordinates") as! NSArray
-                    
-                    
-                    for snapCoordinate in runCoordinates{
-                        let coordinate = snapCoordinate as! NSArray
-                        
-                        let longDict = coordinate[0] as! NSDictionary
-                        let latDict = coordinate[1] as! NSDictionary
-                        
-                            let long:String = longDict.value(forKey: "long") as! String
-                        
-                            let lat:String = latDict.value(forKey: "lat") as! String
-                        
-                            let longCoor = Double(long)
-
-                            let latCoor = Double(lat)
-                            
-                            let runCoordinate = CLLocationCoordinate2D(latitude: latCoor!, longitude: longCoor!)
-                            
-                            coordinateArray.append(runCoordinate)
-                    }
-                    
-                    let pulledRun = Run(uID: uID!, coorArray: coordinateArray)
-                    
-                    self.displayNewShapeWith(newShape: pulledRun.createPulledShape())
-                    self.pulledRunArray.append(pulledRun)
-                }
-            }
-        }
     }
     
     func assignCurrentUser(currentUser: User) {
@@ -225,7 +182,8 @@ class ViewController: UIViewController, GMSMapViewDelegate, LocationManagerDeleg
             
             activeRun.smartArray = activeRun.makeSmartCoordinateArrayfrom(runLocations: activeRun.runLocations)
             
-            let intersectingCoor = activeRun.checkShapeIntersection(existingRuns: pulledRunArray)
+            let intersectingCoor = runManager.checkShapeIntersection(existingRuns: pulledRunArray, activeRun: activeRun)
+            
             for point in intersectingCoor {
                 let marker = GMSMarker(position: point)
                 marker.map = mapView
